07_AUTOMATION_자동화/license_engine/requirements.txt
boto3>=1.34.0
Flask>=3.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.90.0
google-cloud-storage>=2.16.0
WeasyPrint>=60.0
qrcode>=7.4
Pillow>=10.0.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
